+++ CHANGELOG.rst
@@ -3,6 +3,7 @@ Unreleased
 
 Fixed
  * Fixed JSON encoding error with recent FastAPI versions
+ * Fixed ``starlette.HrefMiddleware`` not being compatible with lifespans
 
 Deprecated
  * Python 3.7 is EOL. It is still allowed, but no longer tested.